Use ScopedHandle when possible, fix few tiny warnings.

+++ src/base/filesystem_utils.h
@@ -47,7 +47,7 @@ std::basic_string<CHAR_TYPE> ToLongPathName(
   std::vector<CHAR_TYPE> buffer(src.length() + 1);
   while (true) {
     DWORD api_result = GetLongPathNameImpl(
-        src.c_str(), &buffer[0], buffer.size());
+        src.c_str(), &buffer[0], static_cast<DWORD>(buffer.size()));
     if (api_result > buffer.size()) {
       buffer.resize(api_result);
     }
